requirements.txt
streamlit
python-dotenv
openai
langchain